Added subcommand to template
